llm/requirements.txt
huggingface-hub
llama-cpp-python==0.3.16
mcp==1.20.0
fastapi==0.120.0
uvicorn[standard]
motor
beanie==2.0.0
pydantic
httpx